Bugfixes concerning the DB backup
The password option of mysqldump is used without a space (https://dev.mysql.com/doc/refman/5.7/en/mysqldump.html#option_mysqldump_password).
The database name might differ between installations, it's better to check for the correct name in the config.
The grep code for the DB user didn't work for me, this one does.
IMPORTANT: The grep code for the DB password does not work currently! But as I was unable to provide a quick bugfix for it, I thought I'll start with committing the other fixes.